Year 2023, Day 06 (#36)
Maths for the win.

Today, I realised that brute-forcing the solution (as the text described) wasn't going to be sustainable. I could see that there was going to be a mathematical relationship between the time the button was pressed, and the distance travelled. From that, it was a case of re-discovering how to solve Quadratic Equations and a bit of re-arranging.